Refactor After Merges
Refactored Luke's file parsing code into it's own file and re-wrote main to solve an example case.
