+++ spice-simulator/spice-simulator.vcxproj.filters
@@ -24,12 +24,18 @@
     <ClCompile Include="ConductanceMatrix.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="ParseFile.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Component.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="ConductanceMatrix.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="ParseFile.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
